requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
Flask==3.1.0
fonttools==4.56.0
fpdf2==2.8.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
Werkzeug==3.1.3
